Add robot.pluginsdir property and ROBOT_PLUGINS_DIRECTORY envvar.
In addition to the per-user ~/.robot/plugins directory, search for
plugins in the directory specified by the robot.pluginsdir system
property and the ROBOT_PLUGINS_DIRECTORY environment variable.

If a Jar file with the same name is found in more than one directory,
only the last one that is found will be considered. The per-user
directory is searched last.